+++ src/сomponents/Card/index.jsx
