+++ src/сomponents/FuncButtons/CheckBoxStyled.jsx
