+++ src/сomponents/FuncButtons/Delete.jsx
