+++ src/сomponents/FuncButtons/index.jsx
